FINALLY, added a little bit of tests and lo-and-behold, there are errors in my logic

+++ src/main/java/com/deahstroke/pgcrbatchprocessor/entity/PlayerRaidAggregateStats.java
@@ -32,7 +32,7 @@ public class PlayerRaidAggregateStats {
   private Integer assists;
 
   @Column(name = "hour_played")
-  private Integer hoursPlayed;
+  private Long hoursPlayed;
 
   @Column(name = "clears")
   private Integer clears;
